future mkt framework
This is the first instance of the market system that I plan to add in. Markets will be subsets of the economy with the ability to find prices, and potentially with their own demand and cost weights, likely based on that of the wider imperial system. As for right now, I will set prices in the economy class directly, as I want to reduce the overhead and get better working firm results.